[POM-13] move give-help components corresponding folder
