+++ src/app/app.module.ts
@@ -7,7 +7,7 @@ import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { MatCardModule } from '@angular/material/card';
 import { TranslateLoader, TranslateModule } from "@ngx-translate/core";
 import { PomTranslateLoader, Language } from "./core/translate-loader.service";
-import { AccommodationFormComponentModule } from "./accommodation-form/accommodation-form.module";
+import { AccommodationFormComponentModule } from "./give-help/accommodation-form/accommodation-form.module";
 
 @NgModule({
   declarations: [AppComponent],
