+++ src/app/find-help/find-help.component.html
@@ -1,3 +1,3 @@
-<mat-card style="width: 100%">
+<mat-card>
   <app-accommodation-search></app-accommodation-search>
 </mat-card>
