+++ src/app/give-help/give-help.component.html
@@ -1,0 +1,3 @@
+<mat-card>
+  <app-accommodation-form></app-accommodation-form>
+</mat-card>
