+++ src/app/give-help/give-help.component.scss
@@ -1,0 +1,3 @@
+:host {
+  width: 100%;
+}
