+++ src/app/give-help/give-help.module.ts
@@ -3,15 +3,17 @@ import { CommonModule } from '@angular/common';
 import { GiveHelpComponent } from './give-help.component';
 import { GiveHelpRoutingModule } from "./give-help-routing.module";
 import { MatCardModule } from "@angular/material/card";
+import {AccommodationFormComponentModule} from "./accommodation-form/accommodation-form.module";
 
 @NgModule({
   declarations: [
     GiveHelpComponent
   ],
-  imports: [
-    CommonModule,
-    GiveHelpRoutingModule,
-    MatCardModule,
-  ]
+    imports: [
+        CommonModule,
+        GiveHelpRoutingModule,
+        MatCardModule,
+        AccommodationFormComponentModule,
+    ]
 })
 export class GiveHelpModule { }
