+++ src/app/welcome/welcome.component.html
@@ -1,0 +1,4 @@
+<ul>
+  <li><a routerLink="/daj-pomoc">Daj pomoc</a></li>
+  <li><a routerLink="/szukaj-pomocy">Szykaj pomocy</a></li>
+</ul>
